backend/aparavi-mail-fetcher/aparavi_sdk_example.py
"""
Aparavi SDK Integration Example
This file shows how to integrate with actual Aparavi services when ready.
Currently commented out as it's not integrated.
"""

# EXAMPLE CODE FOR FUTURE INTEGRATION
# Uncomment and configure when ready to integrate with Aparavi

"""
import os
import json
from dotenv import load_dotenv
# from aparavi_dtc_sdk import AparaviClient, PredefinedPipeline

load_dotenv()

class AparaviGmailFetcher:
    '''
    Wrapper class for fetching Gmail data through Aparavi Data Toolchain
    '''
    
    def __init__(self):
        # Initialize Aparavi client
        # self.client = AparaviClient(
        #     base_url=os.getenv("APARAVI_BASE_URL", "https://eaas-dev.aparavi.com"),
        #     api_key=os.getenv("APARAVI_API_KEY")
        # )
        self.pipeline_config_path = os.path.join(
            os.path.dirname(__file__),
            "aparavi-project-export-new-project-2025-10-04T17-32.json"
        )
    
    def validate_pipeline(self):
        '''
        Validate the Gmail pipeline configuration
        '''
        # try:
        #     result = self.client.validate_pipeline(self.pipeline_config_path)
        #     return {
        #         "status": "success",
        #         "message": "Pipeline validation successful",
        #         "data": result
        #     }
        # except Exception as e:
        #     return {
        #         "status": "error",
        #         "message": f"Validation failed: {str(e)}"
        #     }
        pass
    
    def fetch_gmail_emails(self, filters=None, max_results=10):
        '''
        Fetch Gmail emails through Aparavi pipeline
        
        Args:
            filters (dict): Email filters (from, subject, date range, etc.)
            max_results (int): Maximum number of emails to fetch
        
        Returns:
            dict: Fetched emails and metadata
        '''
        # try:
        #     # Execute the pipeline workflow
        #     result = self.client.execute_pipeline_workflow(
        #         pipeline=self.pipeline_config_path,
        #         file_glob="./*.eml"  # Adjust based on your needs
        #     )
        #     
        #     return {
        #         "status": "success",
        #         "data": result,
        #         "filters_applied": filters,
        #         "max_results": max_results
        #     }
        # except Exception as e:
        #     return {
        #         "status": "error",
        #         "message": f"Failed to fetch emails: {str(e)}"
        #     }
        pass
    
    def execute_pipeline_with_monitoring(self):
        '''
        Execute pipeline with full monitoring and status tracking
        '''
        # try:
        #     # Load pipeline config
        #     with open(self.pipeline_config_path, 'r') as f:
        #         pipeline_config = json.load(f)
        #     
        #     # Validate pipeline
        #     validation_result = self.client.validate_pipeline(self.pipeline_config_path)
        #     print(f"Validation: {validation_result}")
        #     
        #     # Execute pipeline
        #     start_result = self.client.execute_pipeline(
        #         pipeline_config,
        #         name="gmail-fetch-task"
        #     )
        #     
        #     if start_result.status == "OK":
        #         token = start_result.data["token"]
        #         task_type = start_result.data["type"]
        #         
        #         # Monitor status
        #         status_result = self.client.get_pipeline_status(
        #             token=token,
        #             task_type=task_type
        #         )
        #         print(f"Status: {status_result}")
        #         
        #         # Cleanup
        #         teardown_result = self.client.teardown_pipeline(
        #             token=token,
        #             task_type=task_type
        #         )
        #         
        #         return {
        #             "status": "success",
        #             "execution": start_result,
        #             "final_status": status_result,
        #             "teardown": teardown_result
        #         }
        #     else:
        #         return {
        #             "status": "error",
        #             "message": "Pipeline execution failed to start"
        #         }
        # 
        # except Exception as e:
        #     return {
        #         "status": "error",
        #         "message": f"Pipeline execution failed: {str(e)}"
        #     }
        pass
    
    def send_webhook_payload(self, webhook_url, payload):
        '''
        Send data to Aparavi webhook for processing
        '''
        # try:
        #     result = self.client.send_payload_to_webhook(
        #         webhook_url=webhook_url,
        #         payload=payload
        #     )
        #     return {
        #         "status": "success",
        #         "data": result
        #     }
        # except Exception as e:
        #     return {
        #         "status": "error",
        #         "message": f"Webhook send failed: {str(e)}"
        #     }
        pass


# Example usage (when integrated):
if __name__ == "__main__":
    print("=" * 60)
    print("Aparavi Gmail Fetcher - SDK Integration Example")
    print("=" * 60)
    print("This is example code for future integration.")
    print("Uncomment the code to use with actual Aparavi SDK.")
    print("=" * 60)
    
    # fetcher = AparaviGmailFetcher()
    # 
    # # Validate pipeline
    # validation = fetcher.validate_pipeline()
    # print(f"Validation: {validation}")
    # 
    # # Fetch emails
    # emails = fetcher.fetch_gmail_emails(
    #     filters={
    #         "from": "claims@insurance.com",
    #         "subject": "claim",
    #         "after": "2024-01-01"
    #     },
    #     max_results=10
    # )
    # print(f"Emails: {emails}")
"""

print("Aparavi SDK Example - Not yet integrated")
print("See commented code for integration examples")
